+++ R/internal.R
@@ -1,3 +1,9 @@
+#' @importFrom forcats fct_reorder
+#' @noRd
+recode <- function(x){
+  as.integer(fct_reorder(x, seq_along(x))) - 1L
+}
+
 #' @importFrom lazyeval f_eval_rhs as.lazy lazy_eval
 #' @importFrom stats terms.formula setNames
 #' @noRd
@@ -27,6 +33,7 @@ getRE2 <- function(data, random, check){
       "Found a random effect with a lone level."
     )
   }
+  #rdat <- lapply(rdat, function(fct) factor(as.integer(fct)))
   RE <- as.data.frame(lapply(setNames(tlabs, tlabs), function(tlab){
     droplevels(lazy_eval(as.lazy(tlab), data = rdat))
   }), check.names = FALSE)
@@ -59,12 +66,8 @@ getZ <- function(RE2){
 
 #' @importFrom stats get_all_vars
 #' @noRd
-getCovariates <- function(data, fixed, random){ # tester fixed = ~1
-  fiFrame <- get_all_vars(fixed, data)
-  if(is.null(random)) random <- ~ 0
-  raFrame <- get_all_vars(random, data)
-  nms <- union(names(fiFrame), names(raFrame))
-  frame <- cbind(fiFrame, raFrame)[nms]
+getCovariates <- function(data, fixed){ 
+  frame <- get_all_vars(fixed, data)
   continuous <- vapply(frame, is.numeric, logical(1L))
   list(
     continuous  = names(frame)[continuous],